Develop

+++ execute_tests.sh
@@ -1,16 +1,25 @@
 #!/bin/bash
 set -e
 
+#Profile checking
+if [ -z "$1" ]; then
+    echo "Error: Profile name missing."
+    echo "Available profiles: smoke-tests, sanity-tests, regression-tests, all-tests"
+    echo "Usage: $0 <profile_id>"
+    exit 1
+fi
+
+PROFILE_NAME=$1
 TIMESTAMP=$(date +"%Y-%m-%d_%H-%M-%S")
 RESULTS_DIR="allure-results"
 REPORT_DIR="allure-report-$TIMESTAMP"
 
-echo "Running tests..."
+echo "Running tests with profile: $PROFILE_NAME"
 
 rm -rf "$RESULTS_DIR"
 mkdir -p "$RESULTS_DIR"
 
-mvn clean test -Dallure.results.directory="$RESULTS_DIR"
+mvn clean test -P="$PROFILE_NAME" -Dallure.results.directory="$RESULTS_DIR"
 
 allure generate "$RESULTS_DIR" -o "$REPORT_DIR" --clean
